+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "baifa",
-  "version": "0.1.0+52",
+  "version": "0.1.0+53",
   "private": true,
   "scripts": {
     "dev": "next dev",
