+++ src/lib/utils/log_parser.ts
@@ -1,4 +1,5 @@
 import { Interface, LogDescription, formatUnits } from 'ethers';
+import { IJsonRpcLog } from '@/interfaces/rpc';
 
 // 定義常見的 ABI 介面
 const COMMON_ABI = [
@@ -45,4 +46,67 @@ export const formatTokenAmount = (rawAmount: string | bigint, decimals: number =
   } catch {
     return '0';
   }
+};
+
+/**
+ * 解析 RPC eth_call 回傳的 ABI 編碼字串 (用於 symbol, name)
+ * @param hex RPC 回傳的 0x 開頭長字串
+ */
+export const parseRpcString = (hex: string | undefined): string => {
+  if (
+    !hex ||
+    hex === '0x' ||
+    hex === '0x0000000000000000000000000000000000000000000000000000000000000000'
+  ) {
+    return '';
+  }
+
+  try {
+    // 1. 移除 0x 前綴
+    const data = hex.startsWith('0x') ? hex.slice(2) : hex;
+
+    /**
+     * ABI String 編碼結構：
+     * [0-63]    - 數據起始位置偏移量 (通常是 0x20)
+     * [64-127]  - 字串實際長度 (以 bytes 為單位)
+     * [128...]  - UTF-8 編碼內容
+     */
+
+    // 2. 取得長度 (位於第 2 個 32-byte 區塊)
+    const lengthHex = data.slice(64, 128);
+    const length = parseInt(lengthHex, 16);
+
+    // 3. 根據長度截取內容 (從第 3 個 32-byte 區塊開始)
+    // 一個 byte 對應兩個 hex 字元，所以從 index 128 開始往後切 length * 2
+    const contentHex = data.slice(128, 128 + length * 2);
+
+    // 4. 將 Hex 轉回 UTF-8 字串
+    const bytes = new Uint8Array(
+      contentHex.match(/.{1,2}/g)?.map((byte) => parseInt(byte, 16)) || []
+    );
+
+    return new TextDecoder().decode(bytes).replace(/\0/g, ''); // 移除可能的空字元
+  } catch (error) {
+    console.error('解析 RPC 字串失敗:', error);
+    return 'Unknown';
+  }
+};
+
+/**
+ * 解析 Logs 並過濾出 ERC20 Transfer
+ */
+export const extractRawTransfers = (logs: IJsonRpcLog[]) => {
+  // ERC20 Transfer 事件的 Topic 0
+  const TRANSFER_TOPIC = '0xddf252ad1be2c89b6c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef';
+
+  return logs
+    .filter(
+      (log) => log.topics[0] === TRANSFER_TOPIC && log.topics.length === 3 // 標準 ERC20 Transfer 會有 3 個 Topic (Event, From, To)
+    )
+    .map((log) => ({
+      tokenAddress: log.address,
+      from: `0x${log.topics[1].slice(26)}`, // 移除補零的部分，還原為 20 bytes 地址
+      to: `0x${log.topics[2].slice(26)}`,
+      value: log.data, // 這是十六進位的金額
+    }));
 };
